More cleanup/refactoring
